+++ src/common.c
@@ -88,21 +88,21 @@ int ishexlc(const char *q)
 	return !*q && !(((p-q))&1);
 }
 
-int ishexn(const char *q, int n)
+int ishexn(const char *q, size_t n)
 {
 	const char *p=q;
 	while (atoi16[ARCH_INDEX(*q)] != 0x7F)
 		++q;
 	return (q-p) >= n;
 }
-int ishexucn(const char *q, int n)
+int ishexucn(const char *q, size_t n)
 {
 	const char *p=q;
 	while (atoi16u[ARCH_INDEX(*q)] != 0x7F)
 		++q;
 	return (q-p) >= n;
 }
-int ishexlcn(const char *q, int n)
+int ishexlcn(const char *q, size_t n)
 {
 	const char *p=q;
 	while (atoi16l[ARCH_INDEX(*q)] != 0x7F)
@@ -134,7 +134,7 @@ int ishexlc_oddOK(const char *q) {
  * the length returned will be length-1 since it would not be proper to try
  * to hex convert the last odd byte.
  */
-static MAYBE_INLINE int _hexlen(const char *q, unsigned char dic[0x100])
+static MAYBE_INLINE size_t _hexlen(const char *q, unsigned char dic[0x100])
 {
 	const char *s = q;
 	size_t len = strlen(q);
@@ -145,15 +145,15 @@ static MAYBE_INLINE int _hexlen(const char *q, unsigned char dic[0x100])
 	if ((size_t)(q - s)&1) --q;
 	return (len == (size_t)(q - s)) ? (int)(q - s) : 0 - (int)(q - s);
 }
-int hexlen(const char *q)
+size_t hexlen(const char *q)
 {
 	return _hexlen(q, atoi16);
 }
-int hexlenu(const char *q)
+size_t hexlenu(const char *q)
 {
 	return _hexlen(q, atoi16u);
 }
-int hexlenl(const char *q)
+size_t hexlenl(const char *q)
 {
 	return _hexlen(q, atoi16l);
 }
@@ -183,13 +183,13 @@ int isdecu(const char *q)
 #ifndef BENCH_BUILD
 /* provides the length of the base64 string.  See base64_convert.c for that
  * function. If the string is not 'pure', then the return is -1*length */
-int base64_mime_len(const char *q) {
+size_t base64_mime_len(const char *q) {
 	return base64_valid_length(q, e_b64_mime, flg_Base64_RET_NEG_IF_NOT_PURE);
 }
-int base64_crypt_len(const char *q) {
+size_t base64_crypt_len(const char *q) {
 	return base64_valid_length(q, e_b64_crypt, flg_Base64_RET_NEG_IF_NOT_PURE);
 }
-int base64_mime_du_len(const char *q) {
+size_t base64_mime_du_len(const char *q) {
 	return base64_valid_length(q, e_b64_mime, flg_Base64_RET_NEG_IF_NOT_PURE|flg_Base64_MIME_DASH_UNDER);
 }
 #endif
